Do not use g_error() to handle expected errors.
g_error() abort()s the program which leads (or may lead, depending on
the user's settings) to a core dump being created. This is absolutely
unnecessary in the case of expected program errors, so use g_critical()
instead and terminate the program normally, with a "bad" exit code.

g_error() is now only used for really critical errors like "Out of
Memory."